Rename metrics.py to logfire.py and move configuration step there
